+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 0.18.0 (2025-05-22)
+
+- Allow to limit VectorStore results by metadata (#564)
+- Switch from imghdr to filetype for image file type check (#563)
+- Remove prompt lab (#549)
+- Add batched() helper method to utils (#555)
+- Rename DocumentMeta create_text_document_from_literal to from_literal (#561)
+- Adjust typing for DocumentSearch (#554)
 - Add Prometheus & Grafana Monitoring for LLMs Using OpenTelemetry (#427)
 - Restructure audit module (#427)